fix: move sole responsibility to save embed stores to MainViewModel to prevent race conditions caused by usage in the onCleared method of multiple ViewModels
